Convert so that all json files have a backup
