+++ HyddwnLauncher/Core/SettingsManager.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Reflection;
 using System.Text;
 using HyddwnLauncher.Extensibility.Interfaces;
+using HyddwnLauncher.Util;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
@@ -67,8 +68,8 @@ namespace HyddwnLauncher.Core
 		{
 			var jsonFile = JsonConvert.SerializeObject(value, type, Formatting.Indented, JsonSerializerSettings);
 
-		    WriteAllTextWithBackup(_configurationFilePath, jsonFile);
-		}
+            jsonFile.WriteAllTextWithBackup(_configurationFilePath);
+        }
 
 		public void SaveSection(Type type, object value)
 		{
@@ -83,14 +84,14 @@ namespace HyddwnLauncher.Core
 				settingsData = JsonConvert.DeserializeObject<dynamic>(jsonFile, JsonSerializerSettings);
 				settingsData[sectionName] = JsonConvert.SerializeObject(value, type, Formatting.Indented, JsonSerializerSettings);
 				jsonFile = JsonConvert.SerializeObject(settingsData, Formatting.Indented);
-			    WriteAllTextWithBackup(_configurationFilePath, jsonFile);
+			    jsonFile.WriteAllTextWithBackup(_configurationFilePath);
                 return;
 			}
 
 			settingsData[sectionName] = JsonConvert.SerializeObject(value, type, Formatting.Indented, JsonSerializerSettings);
 			jsonFile = JsonConvert.SerializeObject(settingsData, Formatting.Indented);
-		    WriteAllTextWithBackup(_configurationFilePath, jsonFile);
-		}
+		    jsonFile.WriteAllTextWithBackup(_configurationFilePath);
+        }
 
 		private class SettingsManagerContractResolver : DefaultContractResolver
 		{
@@ -113,46 +114,5 @@ namespace HyddwnLauncher.Core
 			=> string.IsNullOrWhiteSpace(text)
 				? string.Empty
 				: $"{text[0].ToString().ToLowerInvariant()}{text.Substring(1)}";
-
-
-        /// <summary>
-        /// StackOverflow
-        /// https://stackoverflow.com/questions/25366534/file-writealltext-not-flushing-data-to-disk
-        ///
-        /// When saving the configuration, it ti first written to a temp file.
-        /// If something happens to cause the write to the temp file to fail, the save is lost
-        /// however the original data is untouched. This should reduce loss of configuration data
-        /// due to write failure significantly
-        /// </summary>
-        /// <param name="path"></param>
-        /// <param name="contents"></param>
-	    private static void WriteAllTextWithBackup(string path, string contents)
-	    {
-	        if (!File.Exists(path))
-	        {
-                File.WriteAllText(path, contents);
-	            return;
-	        }
-
-	        // generate a temp filename
-	        var tempPath = Path.GetTempFileName();
-
-	        // create the backup name
-	        var backup = path + ".backup";
-
-	        // delete any existing backups
-	        if (File.Exists(backup))
-	            File.Delete(backup);
-
-	        // get the bytes
-	        var data = Encoding.UTF8.GetBytes(contents);
-
-	        // write the data to a temp file
-	        using (var tempFile = File.Create(tempPath, 4096, FileOptions.WriteThrough))
-	            tempFile.Write(data, 0, data.Length);
-
-	        // replace the contents
-	        File.Replace(tempPath, path, backup);
-	    }
     }
 }
